pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["importers", "ingest.py"]

[project]
name = "family-ledger"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beancount>=3.2.0",
    "fava>=1.26.0",
    "beangulp>=0.1.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[dependency-groups]
dev = [
    "basedpyright>=1.37.0",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.basedpyright]
typeCheckingMode = "standard"
# beangulp does not have type stubs
reportMissingImports = false
